Working towards a better webapp template. Changed the server/app/index system.
new app file to manage the entire web app, services,databases and any other important app structure.
new server file, this handles everything for actaually getting the webserver running.
app ends up being an instance of the server and allows you to pile on top of it.

Soon a nice barebones skeleton will exist that can be employed to get webapp dev started sooner.
This is a major update from my old skeletion that was about 2 years behind in tech.